requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pydantic==2.8.2
requests==2.32.3
httpx==0.27.0
pydantic-settings==2.4.0
orjson==3.10.7
pymysql==1.1.1
cryptography==42.0.8
llama-cpp-python==0.2.56
openai==1.42.0

# Optional model clients (enable as needed)
# google-generativeai==0.7.2
# qianfan==0.3.15  # 文心一言（若采用）
